refactor: convert TypeScript to JavaScript
