+++ README.md
@@ -1,0 +1,7 @@
+# Omora
+
+Omora Chrome extension.
+
+## Development
+
+The project has been refactored to use plain JavaScript modules instead of TypeScript. All former `.ts` files were converted to `.js` and imports updated accordingly.
