--- src/background/index.ts
@@ -1,7 +1,0 @@
-import { loadRegistry } from '../core/registry';
-
-chrome.runtime.onInstalled.addListener(() => {
-});
-
-chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
-});
