+++ src/background/messages.js
@@ -1,0 +1,8 @@
+/**
+ * Message types used within the extension.
+ * @typedef {'activate'|'deactivate'|'resize'|'error'|`feature:event:${string}` } MessageType
+ *
+ * @typedef {Object} Message
+ * @property {MessageType} type
+ * @property {*} [payload]
+ */
