+++ src/background/registry.js
@@ -1,0 +1,1 @@
+export { loadRegistry } from '../core/registry.js';
