--- src/background/registry.ts
@@ -1,1 +1,0 @@
-export { loadRegistry } from '../core/registry';
