+++ src/core/errors.js
@@ -1,5 +1,5 @@
 export class OmoraError extends Error {
-  constructor(message: string) {
+  constructor(message) {
     super(message);
     this.name = 'OmoraError';
   }
