+++ src/core/registry.js
@@ -1,12 +1,10 @@
-import { FeatureMeta } from './types';
-
-export async function loadRegistry(): Promise<FeatureMeta[]> {
-  const metas: FeatureMeta[] = [];
+export async function loadRegistry() {
+  const metas = [];
   try {
     const url = chrome.runtime.getURL('src/features/colorpicker/metadata.json');
     const res = await fetch(url);
     const meta = await res.json();
-    metas.push(meta as FeatureMeta);
+    metas.push(meta);
   } catch (_e) {
   }
   return metas;
