+++ src/core/types.js
@@ -1,0 +1,11 @@
+/**
+ * @typedef {Object} FeatureMeta
+ * @property {string} id
+ * @property {string} name
+ * @property {string} tooltip
+ * @property {string} icon
+ * @property {{script: string, html: string}} entry
+ * @property {string[]} [capabilities]
+ * @property {string} [version]
+ * @property {string} [minOmora]
+ */
