+++ src/ui/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./app.ts"></script>
+    <script type="module" src="./app.js"></script>
   </body>
 </html>
