+++ src/ui/loader.js
@@ -1,6 +1,6 @@
-import { FeatureMeta } from '../core/types';
-import { createSandboxedIframe } from '../core/sandbox';
-export function loadFeature(container: HTMLElement, feature: FeatureMeta): void {
+import { createSandboxedIframe } from '../core/sandbox.js';
+
+export function loadFeature(container, feature) {
   container.innerHTML = '';
   const url = chrome.runtime.getURL(`src/features/${feature.id}/${feature.entry.html}`);
   const iframe = createSandboxedIframe(url);
