+++ src/ui/panel.js
@@ -1,4 +1,4 @@
-export function initPanel(): HTMLElement {
+export function initPanel() {
   const panel = document.createElement('div');
   panel.className = 'om-panel';
   return panel;
